requirements.txt
# Core Dependencies
asyncio==3.4.3
aiohttp==3.9.1
redis==5.0.1
python-dotenv==1.0.0

# Consensus & Communication
grpcio==1.60.0
grpcio-tools==1.60.0

pyzmq==25.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
locust==2.20.0

# Monitoring & Metrics
prometheus-client==0.19.0
psutil==5.9.6

# Utilities
pyyaml==6.0.1
structlog==23.2.0
pydantic==2.5.3

mmh3==4.1.0


cryptography==41.0.7
numpy==1.26.2
scikit-learn==1.3.2